TCK fixes: CMIS-476, CMIS-477, CMIS-478, CMIS-484, CMIS-485, CMIS-486
git-svn-id: https://svn.apache.org/repos/asf/chemistry/opencmis/trunk@1214422 13f79535-47bb-0310-9956-ffa450edef68

+++ chemistry-opencmis-test/chemistry-opencmis-test-tck/src/main/java/org/apache/chemistry/opencmis/tck/tests/crud/UpdateSmokeTest.java
@@ -93,7 +93,9 @@ public class UpdateSmokeTest extends AbstractSessionTest {
             // delete
             deleteObject(doc2);
             if (!doc1.getId().equals(doc2.getId())) {
-                deleteObject(doc1);
+                if (exists(doc1)) {
+                    deleteObject(doc1);
+                }
             }
         } finally {
             // clean up
